Fixed potential issue with integers being silently converted to shorts in the sending/receiving process. Added additional low level logging for tracking down issues with the protocol. Added tests for large payloads. Added sending checks to prevent sending data from overflowing the SendAndReceiveBufferSize.

+++ src/DtronixMessageQueue/MqSession.cs
@@ -94,7 +94,7 @@ namespace DtronixMessageQueue
 
 
             // This will block 
-            Send(buffer, 0, buffer.Length, last);
+            Send(buffer, 0, (ushort)buffer.Length, last);
         }
